requirements.txt
tensorflow>=2.0.0
tf-keras
numpy>=1.19.0
pandas>=1.1.0
matplotlib>=3.3.0
scikit-learn>=0.23.0
pillow>=7.0.0
seaborn
IPython
jinja2